# src/api/swaggerV6.yaml
openapi: 3.0.0
info:
  title: Governify Registry
  description: Governify Registry API
  version: '6'
  termsOfService: www.isa.us.es
  contact:
    name: ISA Research Group
    url: www.isa.us.es
    email: isagroup.us@gmail.com
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
servers:
  - url: /api/v6
    description: Development server
  - url: https://registry.governify.io/api/v6
    description: Production server
...
paths:
...
  /bills/{agreementId}:
    get:
      tags:
        - bills
      summary: Get all bills for an agreement
      operationId: billsGET
      parameters:
        - name: agreementId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bill'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
    
    delete:
      tags:
        - bills
      summary: Remove all bills for an agreement
      operationId: billsDELETE
      parameters:
        - name: agreementId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bills deleted successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
  /bills/{agreementId}/{billId}:
    put:
      tags:
        - bills
      summary: Update a bill's state
      operationId: billsPUT
      parameters:
        - name: agreementId
          in: path
          required: true
          schema:
            type: string
        - name: billId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Bill'
      responses:
        '200':
          description: Bill updated successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
...
components:
  schemas:
...
    ErrorModel:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    Bill:
      type: object
      properties:
        agreementId:
          type: string
        billId:
          type: string
        period:
          ref: '#/components/schemas/Period'
        state:
          type: string
        closeDate:
          type: string
...
    Period:
      type: object
      properties:
        from:
          type: string
        to:
          type: string